Extract service label logic to a re-usable hook
